Add `nerd-icons-ibuffer` package and minor changes
Summary:

This code change adds the `nerd-icons-ibuffer` package to the `init-buffer.el` file. Additionally, there are some minor changes in the `init-python.el` and `init-treesitter.el` files. The `python-shell-completion-native-enable` variable is set to `nil` in `init-python.el`. In `init-treesitter.el`, the `treesit-auto-langs` variable is modified to include additional programming languages.